# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Centralized Knowledge Base API
  description: "API contract for the centralized knowledge base of the Education Center of \nFBUN Central Research Institute of Epidemiology of Rospotrebnadzor.\nDesigned to support parallel backend and frontend development for residency, postgraduate, \nand additional professional education courses.\n"
  version: 1.0.0
servers:
  - url: /api/v1
    description: v1 Base URL
paths:
...
  /documents:
    get:
      summary: List and search documents
      description: Retrieve a paginated list of documents. Supports full-text search, autocompletion/suggestions, and metadata filtering.
      tags:
        - Documents
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Full-text search query supporting synonyms, abbreviations (e.g., 'ФБУН', 'ГЭК', 'ГИА', 'ФГОС'), and auto-suggest.
          required: false
          schema:
            type: string
            example: "ФГОС Эпидемиология"
        - name: suggest
          in: query
          description: Enable autocomplete suggestions flag.
          required: false
          schema:
            type: boolean
            default: false
        - name: doc_type
          in: query
          description: Filter by type of document.
          required: false
          schema:
            $ref: '#/components/schemas/DocumentType'
        - name: specialty
          in: query
          description: Filter by specialty.
          required: false
          schema:
            $ref: '#/components/schemas/Specialty'
        - name: edu_level
          in: query
          description: Filter by education level.
          required: false
          schema:
            $ref: '#/components/schemas/EduLevel'
        - name: category_id
          in: query
          description: Filter by Moodle-aligned category ID.
          required: false
          schema:
            $ref: '#/components/schemas/CategoryId'
        - name: tag
          in: query
          description: Filter by standard tags.
          required: false
          schema:
            type: string
            example: "шаблоны"
        - name: updated_after
          in: query
          description: Filter documents updated after a specific timestamp.
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Page number for pagination.
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Size of pages for pagination.
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: Successful search / retrieval
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  totalElements:
                    type: integer
                    example: 15
                  totalPages:
                    type: integer
                    example: 1
                  pageNumber:
                    type: integer
                    example: 0
                  pageSize:
                    type: integer
                    example: 20
                  suggestions:
                    type: array
                    description: Autocomplete hints or spelling corrections if query is partial.
                    items:
                      type: string
                    example: ["ФГОС ординатура", "ФГОС аспирантура"]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (insufficient roles for a category)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Upload a new document
      description: Upload a file along with metadata properties. Restricted to Content Manager and Admin roles.
      tags:
        - Documents
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - name
                - doc_type
                - specialty
                - edu_level
                - category_id
              properties:
                file:
                  type: string
                  format: binary
                  description: File payload (PDF, DOCX, XLSX, PPTX, JPG, PNG).
                name:
                  type: string
                  description: Document title.
                  example: "Пример рабочей программы ординатуры"
                description:
                  type: string
                  description: Summary of the document.
                  example: "Методические материалы по специальности Эпидемиология."
                doc_type:
                  $ref: '#/components/schemas/DocumentType'
                specialty:
                  $ref: '#/components/schemas/Specialty'
                edu_level:
                  $ref: '#/components/schemas/EduLevel'
                category_id:
                  $ref: '#/components/schemas/CategoryId'
                tags:
                  type: array
                  items:
                    type: string
                  example: ["ординатура", "шаблоны"]
      responses:
        '201':
          description: Document successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid request parameters or file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden for current user role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{id}:
    get:
      summary: Retrieve document details
      description: Get detailed document profile, history metadata, and standard fields.
      tags:
        - Documents
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "9a2fbb22-c35d-4f11-92b1-50e58f00032b"
      responses:
        '200':
          description: Document successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Access control policy violation)
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update document (Create new version)
      description: Atomic update of metadata, description, or replacement of the physical file payload. Automatically tracks incremented version history.
      tags:
        - Documents
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Optional new file payload to increment document version.
                name:
                  type: string
                  example: "Пример рабочей программы ординатуры v2"
                description:
                  type: string
                doc_type:
                  $ref: '#/components/schemas/DocumentType'
                specialty:
                  $ref: '#/components/schemas/Specialty'
                edu_level:
                  $ref: '#/components/schemas/EduLevel'
                category_id:
                  $ref: '#/components/schemas/CategoryId'
                tags:
                  type: array
                  items:
                    type: string
                version_comment:
                  type: string
                  description: Mandatory message describing the update for version history.
                  example: "Обновление ФГОС по приказу Минобрнауки"
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Document not found

    delete:
      summary: Delete a document
      description: Delete a document and its version history. Restricted to Administrators.
      tags:
        - Documents
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document successfully deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Document not found
...
  /audit-logs:
    get:
      summary: Retrieve action logs
      description: View detailed operational history of who performed which actions (creation, modification, deletes) across the platform. Restricted to Admins.
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: category_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CategoryId'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Audit logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogEntry'
                  totalElements:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DocumentType:
      type: string
      enum:
        - Regulations
        - Forms/Templates
        - Protocols
        - Curriculum
        - Guidelines
      description: DB document types category annotations.
    Specialty:
      type: string
      enum:
        - Epidemiology
        - Infectious Diseases
        - Pediatrics
        - Other
    EduLevel:
      type: string
      enum:
        - Residency
        - Postgraduate
        - Additional Professional Education
    CategoryId:
      type: string
      enum:
        - edu_center_root
        - edu_budget_finance
        - edu_staff_workload
        - edu_scholarships
        - edu_academic_reports
      description: Moodle-aligned category path identifiers.
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "9a2fbb22-c35d-4f11-92b1-50e58f00032b"
        name:
          type: string
          example: "Strategic Growth Roadmap"
        description:
          type: string
          example: "Strategic growth goals for FBUN EPIDEM."
        doc_type:
          $ref: '#/components/schemas/DocumentType'
        specialty:
          $ref: '#/components/schemas/Specialty'
        edu_level:
          $ref: '#/components/schemas/EduLevel'
        category_id:
          $ref: '#/components/schemas/CategoryId'
        tags:
          type: array
          items:
            type: string
          example: ["ординатура", "нормативные акты"]
        version:
          type: integer
          example: 3
        fileSize:
          type: integer
          description: Size in bytes
          example: 245000
        fileType:
          type: string
          example: "application/pdf"
        updatedAt:
          type: string
          format: date-time
          example: "2026-08-03T02:29:26Z"
        updatedBy:
          type: string
          example: "ivan.ivanov@epidem.ru"
...
    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: string
          example: "ivan.ivanov@epidem.ru"
        action:
          type: string
          example: "DOCUMENT_UPLOAD"
        resourceId:
          type: string
          example: "9a2fbb22-c35d-4f11-92b1-50e58f00032b"
        category_id:
          $ref: '#/components/schemas/CategoryId'
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "FORBIDDEN"
        message:
          type: string
          example: "Access is strictly denied by default for all financial and personnel categories."